Задача 5

+++ task_3.py
@@ -5,9 +5,9 @@
 # - x=-34; y=-30 -> 3
 
 print('Введите координату x= ', end='')
-x=int(input())
+x=float(input())
 print('Введите координату y= ', end='')
-y=int(input())
+y=float(input())
 
 if x > 0 and y > 0:
     print(f'({x} ; {y}) -> 1 четверть')
